add tests for area charts
